// svelte_webmanager/src/generated/flatbuffers/command.ts
// automatically generated by the FlatBuffers compiler, do not modify

export enum Command {
  NOP = 0,
  START_IAP = 1,
  PING = 2,
  RESET = 3,
  START = 4,
  STOP = 5,
  UP = 6,
  DOWN = 7,
  FORWARD = 8,
  BACKWARD = 9,
  LEFT = 10,
  RIGHT = 11,
  ON = 12,
  ON_FILTER = 13,
  OFF = 14,
  TOGGLE = 15,
  TOGGLE_FILTER = 16,
  TOGGLE_SPECIAL = 17,
  SET_PARAM = 18,
  SET_HORIZONTAL_TARGET = 19,
  STEP_HORIZONTAL = 20,
  SET_VERTICAL_TARGET = 21,
  STEP_VERTICAL = 22,
  SET_LATERAL_TARGET = 23,
  STEP_LATERAL = 24,
  HEARTBEAT = 25,
  SEND_STATUS = 26,
  SET_RGBW = 27,
  SET_SIGNAL = 28,
  SET_PWM = 29,
  DEMO = 30
}